Move assessment helper to shared module
