+++ src/commands/doctor.ts
@@ -4,7 +4,6 @@ import * as Command from "effect/unstable/cli/Command"
 import * as Flag from "effect/unstable/cli/Flag"
 import * as ChildProcessSpawner from "effect/unstable/process/ChildProcessSpawner"
 import type { AssessmentAction } from "#lib/integrations/base.ts"
-import { collectApplicableAssessments } from "#lib/integrations/assessment.ts"
 import knip from "#lib/integrations/tooling/knip.ts"
 import oxfmt from "#lib/integrations/tooling/oxfmt.ts"
 import oxlint from "#lib/integrations/tooling/oxlint.ts"
@@ -14,6 +13,7 @@ import { runMigration } from "#lib/migrations/base.ts"
 import { migrationsById } from "#lib/migrations/index.ts"
 import { DependencyInstaller } from "#lib/services/dependency-installer.ts"
 import { Prompter } from "#lib/services/prompter.ts"
+import { collectApplicableAssessments } from "#lib/shared/assessment.ts"
 import { CommandFailed } from "#lib/shared/errors.ts"
 import { printTitle } from "#lib/shared/terminal.ts"
 import { readPackageJson } from "#lib/workspace/package-json.ts"
