+++ src/commands/update.ts
@@ -4,7 +4,6 @@ import * as Effect from "effect/Effect"
 import * as Command from "effect/unstable/cli/Command"
 import type { ToolingPackage } from "#lib/integrations/base.ts"
 import type { Migration, MigrationCheckResult } from "#lib/migrations/base.ts"
-import { collectApplicableAssessments } from "#lib/integrations/assessment.ts"
 import knip from "#lib/integrations/tooling/knip.ts"
 import oxfmt from "#lib/integrations/tooling/oxfmt.ts"
 import oxlint from "#lib/integrations/tooling/oxlint.ts"
@@ -14,6 +13,7 @@ import { runMigration } from "#lib/migrations/base.ts"
 import { assessmentDrivenMigrationIntegrationById, migrations } from "#lib/migrations/index.ts"
 import { DependencyInstaller } from "#lib/services/dependency-installer.ts"
 import { Prompter } from "#lib/services/prompter.ts"
+import { collectApplicableAssessments } from "#lib/shared/assessment.ts"
 import { printTitle } from "#lib/shared/terminal.ts"
 import { normalizeDependencyVersion, readPackageJson } from "#lib/workspace/package-json.ts"
